Update date-text column tests to pass on all browsers (#1959)
# Pull Request

## 🤨 Rationale

Resolves #1940 

## 👩‍💻 Implementation

Added a page object function to get the expected cell text for a given
value/locale combination. The function internally uses
`Intl.DateTimeFormat` to determine the expected text, so that it will
always return the correct value for a given browser. With this change,
the tests still test all the column, cell, and group row formatting code
without actually testing the nuances of the string returned from
`Intl.DateTimeFormat`.

## 🧪 Testing

Ran the tests for the date-text column in WebKit on my machine (Windows
10) and on the pipeline to verify they pass on both.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.