Disable ts deprecation warnings (#1980)
# Pull Request

## 🤨 Rationale

Disables TypeScript deprecation warnings. Currently we get this warning
in vscode all the time:

![image](https://github.com/ni/nimble/assets/1588923/baaf3622-907a-4422-a3bf-59bb33cb366a)

Which states that a TS feature we have enabled (and will be replaced by
a pretty much compatible feature) will be removed in TS 5.5. We are
pretty far from that because TS 5.5 is not even released yet (let alone
up taken in a version of Angular we support).

We can disable the warning now as the ability to disable warnings was
added in TS 4.9.5 which we recently updated to in #1921

## 👩‍💻 Implementation

Disable TS deprecation warnings with
[ignoreDeprecations](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-0.html#deprecations-and-default-changes).

## 🧪 Testing

Disabled the warning and verified the vscode message is gone. Rely on CI
for the rest.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.