Combobox CVA `addOption` optimization. (#1991)
# Pull Request

## 🤨 Rationale

This PR addresses the root issue behind the actual performance problem
identified in [this AzDO
issue](https://ni.visualstudio.com/DevCentral/_workitems/edit/2686999),
that resulted in reverting a change in `nimble-components` that simply
exposed the CVA flaw.

## 👩‍💻 Implementation

When the CVA adds options, it was always resetting the value of the
`Combobox` in case the newly added option had a value assigned to it
that matched the current assigned model value to the `Combobox`. This
was so that it could update the display text to what the newly added
option had.

However, there is no need to attempt to update the display value if the
newly added option does not have a value that matches the current
`Combobox` value. The fix is to simply put that guard around attempting
to update the display value.

## 🧪 Testing

Unit tests still pass. Manual testing, and ran performance checks
against a re-implementation of the `ListOptionOwner` by the
`nimble-components` `Combobox`, and verified there was no slowdown.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>