Created a new Experimental Prerendering module (#1955)
# Pull Request

## 🤨 Rationale

Created a new experimental prerendering module.

The previous prerendering oversaw calculating the `fillstyle` of each
die and it's label.
We want to move these expensive operations to the web worker, and to
make them faster.

## 👩‍💻 Implementation

I removed the code which prepared the dies for rendering.
I created a domain splitter for linear color scales. this will reduce
the number of colors displayed but will increase the performance because
we are no longer finding interpolations for each individual value in the
whole domain.

Open Issues:
- given the problems with arquero, should we implement in-house
highlight or leave it unimplemented until the issue gets fixed?

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation. If any functionality is
not covered by automated testing, provide justification.
-->

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>