Fork FAST Combobox. (#1973)
# Pull Request

## 🤨 Rationale

- #1948 

As part of the effort to bring the `Combobox` implementation in
alignment with the `Select`, namely with how we handle controlling the
visibility of the options in the dropdown, as well as the need to
register options as part of the `connectedCallback` hook, it makes sense
to fork the FAST `Combobox` implementation into our own, as we did with
the `Select`, to make it more clear what is happening within our own
implementation for the various new behaviors we require.

## 👩‍💻 Implementation

Essentially bringing over the current `Combobox` implementation from the
FAST `archives/fast-element1` branch into our own. Copied the FAST
`FormAssociatedCombobox` that our `Combobox` now extends.

## 🧪 Testing

Ran existing unit tests. Manual testing. Verified that our existing
tests still replicate all of the tests currently in the FAST
`archives/fast-element1` repo.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>