Update issue referenced in skipped tests (#2000)
# Pull Request

## 🤨 Rationale

The tests that were marked as being skipped in Firefox and Webkit
because of #1943 are failing in those browsers for the same reason
described in more detail in #1936. Therefore, I'm updating the comments
to reference #1936 and closing #1943 as a duplicate.

Closes #1943

## 👩‍💻 Implementation

Update comment to point to different issue

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.