Move column validator into columnInternals (#1999)
# Pull Request

## 🤨 Rationale

Resolves #1389

## 👩‍💻 Implementation

- Updated `TableColumn`, `ColumnInternals`, and `ColumnInternalsOptions`
to be generic with `TColumnValidator extends ColumnValidator<[]>`
- Updated `ColumnInternals` to have a `validator` property whose type is
enforced by the type of `TColumnValidator`
- Made `validator` a required input on `ColumnInternalsOptions`, so now
every column has to specify a validator. In some cases this is a
"default" validator with no validity keys, but making this required
simplified a few different code paths. This included keeping columns
with validators from having to use the `!` operator to assert that they
knew that a validator had been set within their options.
- Updated columns to specify a validator and to use that validator
through `ColumnInternals` rather than a class member on the column.

## 🧪 Testing

Existing tests still pass

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.