Fix and re-enable intermittent mention-listbox test (#2006)
# Pull Request

## 🤨 Rationale

Fixes #1891 

It's not clear why, but the mention-listbox test was awaiting
`waitForUpdatesAsync` twice before asserting that the scroll position
was correct. Other similar tests await `waitForUpdatesAsync` followed by
`waitAnimationFrame`. The call to `waitAnimationFrame` is necessary
because the scroll is queued with `requestAnimationFrame`.

## 👩‍💻 Implementation

Update the second call to `waitForUpdatesAsync` to a call to
`waitAnimationFrame`.

## 🧪 Testing

Ran the test a number of times locally, though I'm not if this verifies
much since I was never able to get the test to fail locally before this
fix

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.