Add configuration for icon mappings and spinner mappings to show/hide their text within a table's cells (#2012)
# Pull Request

## 🤨 Rationale

This updates the `nimble-mapping-icon` and `nimble-mapping-spinner`
elements to be aligned with the changes described in #1992.

## 👩‍💻 Implementation

- Added `text-hidden` to `nimble-mapping-icon` and
`nimble-mapping-spinner`
- I also added these to Blazor and Angular so that when the change is
published clients in any framework can uptake the new package and have
the opportunity to add `text-hidden` to mappings in order to keep their
existing behavior
- Reworked the `MappingIconConfig` and `MappingSpinnerConfig` to be more
aligned
- They both create the appropriate `ViewTemplate` to render a spinner,
icon, or blank
- They both add the appropriate `title` and `aria-*` attributes based on
the configuration
- Neither adds css classes that are styled in the group header styles or
cell view styles
- Update the templates for the icon column's cell view and group header
share code for rendering icons and spinners (enabled because of aligning
`MappingIconConfig` and `MappingSpinnerConfig`) and to render the text
based on the `text-hidden` attribute

## 🧪 Testing

- Updated storybook & matrix tests
- Manually tested
- Wrote/updated a number of tests for the icon column
- To enable this, created a page object for the icon column to
encapsulate the logic around finding the icon/spinner and the text

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.