Support text mappings in icon column (#2023)
# Pull Request

## 🤨 Rationale

This adds support to `nimble-table-column-icon`* for text mappings as
described in #1992.

*`nimble-table-column-icon` will be renamed to
`nimble-table-column-mapping` in a future PR

## 👩‍💻 Implementation

I've updated the column validator to allow text mappings and updated the
cell-view and group-header-view to render text mappings correctly.

## 🧪 Testing

- Updated storybook and matrix tests
- New unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.