Don't allow use of the nullish coalescing operator in templates (#2025)
# Pull Request

## 🤨 Rationale

Resolves #1843 

I did an evaluation of how observables work within FAST templates. A
longer explanation of what is and is not allowed can be found [in the
discussion within the linked
issue](https://github.com/ni/nimble/issues/1843#issuecomment-2064663561).
The only issue I found within nimble code was the usage of `??` in
templates, which may not correctly be identified as a volatile binding.
Therefore, I've rewritten the bindings that previously used `??` and
added a lint rule for `template.ts` files that disallows usage of `??`.

## 👩‍💻 Implementation

See Rationale

## 🧪 Testing

All existing tests still pass
Verified that the new lint rule catches instances of `??` in
`template.ts` files

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.