Angular export mapping key type (#2032)
# Pull Request

## 🤨 Rationale

In #2026, the `enum-text` column was removed from nimble. However, I
hadn't realized that in Angular, `NimbleTableColumnEnumTextModule` was
the only module that exported `MappingKeyType`, which was expected to be
used by both the `enum-text` column and the `icon` column. Therefore,
the `icon` column is now unusable in Angular because `MappingKeyType`
isn't exported.

Therefore, this PR exports `MappingKeyType` from
`NimbleTableColumnIconModule`.

## 👩‍💻 Implementation

Export `MappingKeyType` from `NimbleTableColumnIconModule`

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.