Include Fallback font in font and font-family tokens (#2033)
# Pull Request

## 🤨 Rationale

Found that the nimble landing page was not using the fallback fonts when
loading on a slow network. The reason was that the fallback fonts were
previously not included in the font-family token but were a separate
token:
https://github.com/ni/nimble/blob/bad9f4c573bca5528355ec93e10bafe12b76e386/packages/site/landing/styles.scss#L110

This change updates the `font-family` theme tokens to always be defined
as the base token font family combined with the fallback font family.
This impacts the definition of the `font-family` tokens (but `font` is
effectively unchanged). From [codesearch for font-family
usage](https://codesearch.natinst.com/search/text?q=nimble.*font.*family&fold_case=auto&regex=true&context=true)
this will impact those usages but hopefully for the better via improved
fallback fonts during page load.

In addition it removes the fallback font tokens from being explicitly
exported to prevent accidental direct usage. I found one direct usage of
the [fallback token in
codesearch](https://codesearch.natinst.com/search/text?q=nimble.*fallback.*family&fold_case=auto&regex=true&context=true)
but looks like they are leveraging the SCSS build so it should be
identified as part of build in an upgrade.

Partially addresses https://github.com/ni/nimble/issues/1192

## 👩‍💻 Implementation

Went for defining the Fallback font by convention. It assumes that all
our Font Families that are defined will also define a Fallback using a
known convention of `[Font Family] Fallback`.

Alternatives: Create a bunch of tokens for the Fallback font families as
initially discussed in the issue. It didn't really seem high value. The
definition of the fallback fonts is completely encapsulated in the CSS
files using CSS features. Outside the web context having the token for
the fallback font is not useful unless we also encode all the state for
the fallback (size-adjust, etc) and even that might be CSS-specific. And
in the web context you shouldn't really use it directly and rely on the
theme token.

## 🧪 Testing

Rely on CI.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed. I think that the font README
sufficiently covers the need for creating Fallback tokens.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>