Add Storybook for `spright-components` (#2030)
## 🤨 Rationale

Part of #1978

## 👩‍💻 Implementation

Long-term, we want a single Storybook for both Nimble and Spright. This
will mean pulling all the Storybook code and story files into a separate
package that depends on `nimble-components` and `spright-components`.
Initially, we'll create this separate package, but it will only contain
the Spright stories. Later, we will move all of Nimble's stories into
that package and start publishing from the new package instead of the
Storybook in `nimble-components`.

- Created `storybook` with:
   - `.storybook` directory (copied from `nimble-components`)
- a handful of Storybook-related utility files (copied from
`nimble-components/src/utilities/tests`)
   - React wrappers for check and exclamation mark icons used by stories
   - standard story and visual tests for Spright Rectangle component
- Removed Storybook-related dependencies and configuration from
`spright-components`

## 🧪 Testing

Built and ran Storybook locally.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>