Allow icon column to be configured as fixed width (#1907)
# Pull Request

## 🤨 Rationale

This change is part of #1873 and follows the decisions made in #1992

This change adds the ability to configure the `width-mode` of the icon
column. When set to `iconSize`, the column is a fixed width of 32px,
cannot be resized, and hides its sorting & grouping indicators in the
header.

## 👩‍💻 Implementation

There were a number of changes that needed to be made within nimble to
facilitate this change.

1. Column Internals
- Add boolean properties of `resizingDisabled` and
`hideHeaderIndicators`
1. Table Update Tracker
- Mark `columnWidths` as being changed when `resizingDisabled` changes
in `columnInternals`
1. Table Layout Manager
    - Update to deal with columns that have `resizingDisabled === true`
1. Table
- Add logic to hide column dividers that can never be dragged to the
left or to the right. This is accomplished by adding a `draggable` class
to the dividers that can be dragged and making the presence of that
class a condition in the CSS for showing the divider.
- Add logic to set `indicators-hidden` on table columns where
`hideHeaderIndicators === true`
1. Table Header
- Add a boolean attribute of `indicators-hidden` that, when `true`,
hides the sorting & grouping indicators in the column header.
1. Icon Column
- Add `width-mode` attribute. When the attribute is changed,
`columnInternals` is updated with appropriate values for
`resizingDisabled`, `hideHeaderIndicators`, `minPixelWidth`, and
`pixelWidth`.

## 🧪 Testing

Manually tested in storybook
Updated & added unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>