Move anchored region import from menu to menu item (#2051)
# Pull Request

## 🤨 Rationale

Saw an issue in the menu matrix story where sub-menus looked wrong.
Turns out it was because `AnchoredRegion` had not been registered, and
FAST's menu item template was using `tagFor(AnchoredRegion)`.

## 👩‍💻 Implementation

We already had an import of `anchored-region` in the menu code, but it
really belongs in menu item. Moved it.

## 🧪 Testing

Trusting Chromatic.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.