Blazor and Spright invoke publish fix (#2059)
# Pull Request

## 🤨 Rationale

Ran the `invoke-publish:setup` command locally and found that it was not
behaving as expected. Old packages were not removed. Realized the
commands were running in the wrong directory.

## 👩‍💻 Implementation

Update the `invoke-publish:setup` and `invoke-publish:nuget` commands
for Nimble and Spright to point to the correct directory.

## 🧪 Testing

Validated `invoke-publish:setup` locally and relying on publish in CI
for the package publish 🤞

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.