Anchor menu item focus styling fix (#2050)
# Pull Request

## 🤨 Rationale

Fixes #2028

## 👩‍💻 Implementation

The root of the issue is that the focus outline was being shown any time
focus was within the anchor menu item rather than being styled based on
focus-visible. The difficulty is that the anchor menu item delegates
focus to the inner anchor element, which means the anchor becomes
focus-visible but the anchor menu item does not. Therefore, I reworked
the styling a little bit so that the anchor element is the same size as
the anchor menu item to allow the anchor to have the outline. A
beneficial side effect of this change is that the click area of the
anchor now exactly matches the hover & focus area of the menu item.

## 🧪 Testing

- Updated the menu button story to include an anchor menu item as the
first item to verify the behavior
- Verified no chromatic changes were introduced

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>