Storybook CONTRIBUTING doc improvements (#2075)
# Pull Request

## 🤨 Rationale

#2044 removed instructions for contributors about what Storybook files
need to be created for each component. I think that info is valuable for
external contributors so I want to revive it in an appropriate place.

## 👩‍💻 Implementation

1. Revived the docs that existed before in
nimble-components/CONTRIBUTING.md and put them in
storybook/CONTRIBUTING.md and added additonal context
2. Linked to those docs from nimble-components/CONTRIBUTING.md
3. Updated storybook README contributing section
4. Other minor tweaks to reflect new patterns and commands

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [X] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: m-akinc <7282195+m-akinc@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>