Refactor common Blazor acceptance test code (#2074)
# Pull Request

## 🤨 Rationale

Part of #1976

## 👩‍💻 Implementation

I've created a new project in `blazor-workspace` to contain code shared
between the Nimble and Spright acceptance test projects. I ran into
issues with the page components not being found when running tests from
the VS Test Explorer. Through trial and error, I discovered that the
`Setup` instance needed to be created in the same assembly as the page
components for them to be discovered. Normally this is governed by the
`AppAssembly` property of the `Router` component (in `App.razor`), but
that is not used by the test runner.

## 🧪 Testing

Tests pass.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>