Create nimble-mapping-empty component (#2077)
# Pull Request

## 🤨 Rationale

As part of #1992, we decided to create a `nimble-mapping-empty`
component that would allow a cell to render as blank for a given mapping
while still having text provided for use in group rows.

This PR creates that component within `nimble-components`. Support for
Blazor and Angular will be added in a subsequent PR.

## 👩‍💻 Implementation

- Create `nimble-mapping-empty` component
- Add support for the component in the mapping column
- Update tests

## 🧪 Testing

- Extended the unit tests for the mapping column to include tests for
empty mappings
- Add empty mapping to the matrix test for the mapping column

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.