Require an icon to be specified on icon mappings (#2103)
# Pull Request

## 🤨 Rationale

In #1874, we introduced the ability to have icon mappings with an
`undefined` icon to support use cases where the cell should be empty
with non-empty group rows and where the cell should have only text.
These use cases are now covered by usages of `nimble-mapping-empty` and
`nimble-mapping-text`, both of which are supported by the mapping
column. Therefore, we no longer need to support icon mappings without an
icon.

## 👩‍💻 Implementation

Mostly reverting #1874, with a few minor updates to account for updates
made in the source since the change was put it.

## 🧪 Testing

- Added new unit test that verifies setting an `undefined` icon on an
icon mapping in the mapping column results in a validation error
- Updated matrix tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.