Use display helper consistently for all components (#2114)
# Pull Request

## 🤨 Rationale

This PR updates all defined components to use a custom version of the
`display` helper. The goal of this PR is to be a refactor to use the
same helper but should be effectively a no-op.

Part of #1326

## 👩‍💻 Implementation

- Added a custom display helper
- Updated all elements to use it
- Updates elements that were missing it to have one with a value that
reflects what shows up in storybook today as the computed display value
- Utility elements that should never be visible / have visible content
were set to `display: none;`

## 🧪 Testing

Rely on storybook to verify nothing changed.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed. Updated the `display` doc topic.