Fix text field start slot color (#2108)
# Pull Request

## 🤨 Rationale

An icon put in the `start` slot of the `nimble-text-field` had the
incorrect color based on the visual design spec. Specifically, the icon
had been the text color, but it should have a 60% opacity applied to it.

## 👩‍💻 Implementation

Updated the styling of the `start` slot

## 🧪 Testing

Manually tested across all 3 themes in storybook

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.