Angular i18n enable (#2122)
# Pull Request

## 🤨 Rationale

Enables the i18n attribute lint rule for the angular-workspace. The goal
is to have some level of automation in detecting attributes that should
not be marked i18n early to add to the javascript style guide angular
eslint configuration.

## 👩‍💻 Implementation

- Enabled the rule with a local list of exceptions.
- Found that even though the i18n attributes rule is enabled for all
files, it does not seem to run for our tests. Seems to be because our
inline templates in tests are not defined at the top-level of the file.
Created angular-eslint issue:
https://github.com/angular-eslint/angular-eslint/issues/1825

## 🧪 Testing

Rely on CI.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed. Comments are in the eslint config,
didn't seem like additional were neecessary.