Rebuild lock file and update dependencies (#2123)
# Pull Request

## 🤨 Rationale

It was identified in #2121 that the angular refactor resulted in
leftover entries in the lock file. The entries were marked `extraneous`
and did not seem to interfere with the build, etc. This PR rebuilds the
lock file which cleans out extraneous entries and updates dependencies
to package.json supported range latest.

## 👩‍💻 Implementation

```
git clean -fdx
del package-lock.json
npm install
```

## 🧪 Testing

Rely on CI.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.