Improve API docs for anchor and tab components (#2126)
# Pull Request

## 🤨 Rationale

As part of #824 we want to improve the accuracy and comprehensiveness of
our storybook API documentation. This PR covers:
- anchor
- anchor button
- anchor tab and related components
- tab and related components

## 👩‍💻 Implementation

1. Followed same patterns as previous PRs to restructure MDX files and
to add API categories to story files
2. Refactored anchor tab and tab to include separate stories for their
child components, similar to what I did for menu

## 🧪 Testing

Storybook exploration

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.