Storybook link directly to components source code (#2131)
# Pull Request

## 🤨 Rationale

This PR is an abomination but hopefully fixes
https://github.com/ni/nimble/issues/2118

Turbosnap relies on git source history to detect changes. As a separate
package, the storybook build was correctly implemented to rely on the
build output of nimble-components and spright-components instead of the
non-public source.

However there is not a link that Turbosnap can detect from the source in
git to the ephemeral dist output paths that ends up being used during
build. So Turbosnap was now unable to follow the dependency path from a
story file to specific component source files.

## 👩‍💻 Implementation

Do as minimally scoped a change as possible to enable storybook to link
directly to non-public source code in the nimble-components and
spright-components packages. Any other non-mechanical changes have
comments in the PR.

## 🧪 Testing

I think [this
build](https://github.com/ni/nimble/actions/runs/9194886405/job/25289396570#step:16:78)
where I reverted a change that was only in the spright rectangle css
source confirms that the approach will work. You can see turbosnap
detected the original component source file changes for the rectangle
story in the log:

```
Found 5 changed files:
  change/@ni-spright-components-a39f0ecb-4139-47f6-86f7-c22779ee1f08.json
  package.json
  packages/spright-components/src/rectangle/styles.ts
  packages/storybook/.eslintrc.cjs
  packages/storybook/README.md
...
Found affected story files:
  packages/storybook/src/spright/rectangle/rectangle-matrix.stories.ts [./src/spright/rectangle/rectangle-matrix.stories.ts]
  packages/storybook/src/spright/rectangle/rectangle.stories.ts [./src/spright/rectangle/rectangle.stories.ts]
  packages/storybook/src/spright/rectangle/rectangle.mdx [./src/spright/rectangle/rectangle.mdx]
...
✔ TurboSnap enabled
Capturing 8 snapshots and skipping 582 snapshots.
```

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed. Updated the storybook README with a
warning this is not a general pattern to copy.