Update banner button styles (#2130)
# Pull Request

## 🤨 Rationale

The "dismiss" button in the `nimble-banner` had the wrong styling in the
following ways:
- It was the wrong size. It was 16x16, but it should have been 24x24.
- It was using the wrong icon. It was using the `xmark` icon, but it
should have been using the `times` icon.
- The hover styling was wrong. It had custom styling to make it look
like a ghost button with a fill on hover (not like any of our standard
`nimble-button` appearance modes). Instead, it should be a standard
`content-hidden`, `appearance="ghost"`, `appearance-variant="primary"`
button.

Note: A number of these differences were because the design spec was
updated after the banner was initially implemented.

## 👩‍💻 Implementation

1. Wrapped the contents of the banner's template in a theme provider
hard-coded to the `color` theme
2. Updated storybook docs to specify that a slotted `action` button
should use the `primary` appearance variant
3. Updated instances of the banner within nimble to set
`appearance-variant="primary"` on slotted buttons

Those three changes allowed me to remove a lot of custom styling from
the banner.

## 🧪 Testing

- Verified that the only chromatic change was due to the "dismiss"
button

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.